Typo: missing line in configs "short syntax" example
The paragraph above says "grant the `redis` service access to the
`my_config` and `my_other_config` configs."  While `my_other_config`
is defined, it is not added to the `redis` `configs` list.

I suspect this was a merge error or something.

To ensure I wasn't going crazy, I verified that at least Docker
Compose v2 requires this line for the second config to be
materialized inside the service.

Signed-off-by: Jay Carlson <nop@nop.com>