schema: remove unrecognized keywords
I noticed in passing that some of the entries inside the JSON Schema did
not do anything:
- under 2019-09 `id` does not mean anything (as of draft-06, it's
[spelled `$id`](https://json-schema.org/draft-06/json-schema-release-notes)), and
there's no need to define explicit IDs for internal references.
- there is no recongized "expose" format
- the `constraints` definition is never referenced and has no constraints
(`service` is not a valid keyword).

I've chosen a top level `$id` URI that works, but this is arbitrary and
could be replaced by something less github-specific: although not a
requirement, it's nice if this is a working URL.

I've also added a newline at the end of the file because that's
reasonably standard.

This new schema should be functionally identical to the original.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>